refactor: an email is named after the address, not the role
HELLO_EMAIL leaves room for the next one. install.ts becomes cli.ts, since what
it holds is the cli's install rather than any install.
