+++ packages/global-constants/src/cli.ts
